session management and logout feature added

+++ springapp/src/main/java/com/example/springapp/controller/MovieController.java
@@ -59,8 +59,6 @@ public class MovieController {
 	@Autowired
 	private WorkedOnService workedOnService;// Object to connect to workedOn service of service layer
 
-	// @Autowired
-	// private MovieRepository movieRepository;
 	
 	// object to connect to JWT configuration
 	@Autowired
@@ -113,20 +111,44 @@ public class MovieController {
 		try {
 			User user = this.userService.AuthenticateUser(email, password);
 			if (user != null) {
-				String token = jwtTokenUtil.generateToken(user);
+				String token = user.getJwtToken();
+				if(token != null && !jwtTokenUtil.isTokenExpired(token)) {
+					return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+				}else {
+					userService.saveToken(user, null);
+				}
+				token = jwtTokenUtil.generateToken(user);
 				Map<String, Object> responseBody = new HashMap<>();
 				responseBody.put("token", token);
 				responseBody.put("userId", user.getUserId());
 				responseBody.put("email", user.getEmail());
 				responseBody.put("name", user.getName());
 				responseBody.put("role", user.getRole());
+				userService.saveToken(user, token);
 				return ResponseEntity.ok(responseBody);
 			}
-			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		} catch (Exception e) {
 			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
 		}
 
+	}
+	
+	@PostMapping("/signout")
+	public ResponseEntity<HttpStatus> signOut(@RequestHeader(name = "Authorization") String token,@RequestParam("email") String emai){
+		token = token.substring(7);
+		String email;
+		try {
+			email = jwtTokenUtil.getUsernameFromToken(token);
+			User user = userService.getUserByEmail(email);
+			userService.saveToken(user, null);
+			return new ResponseEntity<>(HttpStatus.OK);
+		}catch (ExpiredJwtException e) {
+			userService.replaceToken(token, null);
+			return new ResponseEntity<>(HttpStatus.ACCEPTED);
+		}catch(Exception e) {
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 	//check if a token is valid for a particular user
@@ -163,7 +185,8 @@ public class MovieController {
 			if (token != null && token.startsWith("Bearer ")) {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
-					if (user.getEmail().equals(jwtTokenUtil.getUsernameFromToken(token))) {
+					String email = jwtTokenUtil.getUsernameFromToken(token);
+					if (user.getEmail().equals(email) && userService.authenticateToken(email, token)!= null) {
 						responseBody.put("userId", user.getUserId());
 						responseBody.put("email", user.getEmail());
 						responseBody.put("name", user.getName());
@@ -205,43 +228,7 @@ public class MovieController {
 	}
 
 	
-	// @GetMapping("/movies/highestrated")
-	// public ResponseEntity<List<Movie>> getHighestRatedMovies(){
-	// 	try{
-	// 		List<Movie> movies = this.movieService.getHighestRatedMovies();
-	// 		return ResponseEntity.status(HttpStatus.OK).body(movies);
-	// 	}catch(Exception e){
-	// 		return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-	// 	}
-	// }
 
-
-	// @GetMapping("/movies/recent")
-	// public ResponseEntity<List<Movie>> getRecentMovies(){
-	// 	try{
-	// 		List<Movie> movies = this.movieService.getRecentMovies();
-	// 		return ResponseEntity.status(HttpStatus.OK).body(movies);
-	// 	}catch(Exception e){
-	// 		return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-	// 	}
-	// }
-	
-
-	// retrieves the particular movie data that has primary key as movieId
-	// @GetMapping("/movies/{movieId}")
-	// public ResponseEntity<MovieModel> getMovie(@PathVariable String movieId) {
-	// 	try {
-	// 		Movie m = this.movieService.getMovie(Long.parseLong(movieId));
-	// 		if (m != null) {
-	// 			List<WorkedOn> wl = workedOnService.getList(m);
-	// 			return ResponseEntity.status(HttpStatus.OK).body(new MovieModel(m, wl));
-	// 		} else {
-	// 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-	// 		}
-	// 	} catch (Exception e) {
-	// 		return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-	// 	}
-	// }
 
 	// creates a new movie object and stores the data in the database after
 	// authenticating the user as admin
@@ -257,7 +244,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						Movie movie = new Movie();
 						movie.setTitle(title);
 						movie.setGenre(genre);
@@ -300,7 +287,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						Movie movie = new Movie();
 						movie.setMovieId(Long.parseLong(movieId));
 						movie.setTitle(title);
@@ -335,7 +322,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						this.movieService.deleteMovie(Long.parseLong(movieId));
 						return new ResponseEntity<>(HttpStatus.OK);
 					}
@@ -395,7 +382,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null) {
 						Review review = new Review();
 						review.setReviewNote(reviewNote);
 						review.setRating(rating);
@@ -440,7 +427,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						Review review = reviewService.updateReview(Long.parseLong(reviewId), reviewNote, rating, source);
 						if (review != null) {
 							Movie movie = review.getMovie();
@@ -474,7 +461,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						Movie movie = reviewService.deleteReview(Long.parseLong(reviewId));
 						if (movie != null) {
 							String rating = reviewService.getRating(movie);
@@ -540,7 +527,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 
 						String filename = (poster != null) ? handleFile(poster) : null;
 						if (castService.addCast(name, filename) != null) {
@@ -568,7 +555,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 
 						Cast cast = castService.getCast(castId);
 						Movie movie = movieService.getMovie(movieId);
@@ -600,7 +587,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						Cast cast = castService.getCast(Long.parseLong(castId));
 						if (cast != null) {
 							workedOnService.deleteByCast(cast);
@@ -630,7 +617,7 @@ public class MovieController {
 				token = token.substring(7);
 				if (!jwtTokenUtil.isTokenExpired(token)) {
 					String email = jwtTokenUtil.getUsernameFromToken(token);
-					if (userService.isEmailExist(email) && userService.isUserAdmin(email)) {
+					if (userService.isEmailExist(email) && userService.authenticateToken(email, token)!= null && userService.isUserAdmin(email)) {
 						Cast cast = castService.getCast(Long.parseLong(castId));
 						Movie movie = movieService.getMovie(Long.parseLong(movieId));
 						if (cast != null && movie != null) {
